Add SVGView
Thought that SVG could be a separate category. I saw a library here that converts SVG images to UIImages in image category, but this library converts every SVG element to SwiftUI elements, so I didn't find a category for it. And seeing as PDF has its own category I decided to go with a separate category for SVG as well. Also contribution guide says to create issue in case I want new category, but should I open a PR in that case as well or just create the issue? I'll create one just in case. Have a great day!